% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLogExp.R
\name{avgLogExp}
\alias{avgLogExp}
\title{Compute average log expression}
\usage{
avgLogExp(object, features = NULL, cells = NULL, pseudocount.use = 1)
}
\value{
Mean in log space
}
\description{
Custom function to compute log average expression level
}
