% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedViolin.R
\name{extract_max}
\alias{extract_max}
\title{Extract max value of y axis}
\usage{
extract_max(p)
}
\arguments{
\item{obj}{A Seurat object.}

\item{features}{Genes to be plotted.}

\item{pt.size}{Size of jitter points. Default is 0 (No point is plotted).}

\item{plot.margin}{Adjust the white space between each plot.}
}
\value{
ymax
}
\description{
Helper function to reproduce scanpy stacked violin plot with Seurat.
}
\details{
Credit to: Tang Ming
https://divingintogeneticsandgenomics.rbind.io/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/
}
