% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosketch.R
\name{geosketch}
\alias{geosketch}
\title{Geometric Sketching}
\usage{
geosketch(object, sketch.size = 1000, assay = "RNA", slot = "data", k = 20)
}
\arguments{
\item{object}{A Seurat object.}

\item{sketch.size}{Number of cells to sketch. By default sketch 1000 cells.}

\item{assay}{Assay used for sketching. "RNA" by default.}

\item{slot}{By default slot = "data".}

\item{k}{Target rank of the low-rank SVD. k should satisfy k << min(m,n). By default k = 20.}
}
\value{
A sketched Seurat object.
}
\description{
Geometric Sketching to subsample data from single cell genomics into smaller representative subset of cells, while preserving
biological complexity, highlighting rare cell states, and accelerating complex analysis such as data integration.
}
\details{
The original package `geosketch` is written in Python. With `reticulate`, we implement it in R and provides a function that directly works with
the Seurat object.

Publication: Hie, B., Cho, H., DeMeo, B., Bryson, B., and Berger, B. (2019). Geometric Sketching Compactly Summarizes the Single-Cell
Transcriptomic Landscape. Cell Systems.

Github repo: https://github.com/brianhie/geosketch
}
\examples{
## If haven't installed already.
reticulate::py_install("geosketch")

## mgc has 7527 cells.
mgc.sketched <- geosketch(mgc)

}
