% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedViolin.R
\name{modify_vlnplot}
\alias{modify_vlnplot}
\title{Parse Seurat violin plot}
\usage{
modify_vlnplot(
  obj,
  feature,
  pt.size = 0,
  plot.margin = ggplot2::unit(c(-0.75, 0, -0.75, 0), "cm"),
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{pt.size}{Size of jitter points. Default is 0 (No point is plotted).}

\item{plot.margin}{Adjust the white space between each plot.}

\item{features}{Genes to be plotted.}
}
\value{
A ggplot object
}
\description{
Helper function to reproduce scanpy stacked violin plot with Seurat.
}
\details{
With violin plot generated by `Seurat::VlnPlot`, this function removes
its x-axis text and tick, and adjusts the white space between each
plot with plot.margin argument.
... pass any arguments to VlnPlot in Seurat
Credit to: Tang Ming
https://divingintogeneticsandgenomics.rbind.io/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/
}
