% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedViolin.R
\name{parse_title}
\alias{parse_title}
\title{Parse title length}
\usage{
parse_title(wrap_title, break_length = 23)
}
\arguments{
\item{wrap_title}{Title to be plotted.}

\item{break_length}{Length of break point: where to insert line break?}
}
\value{
String of title with appropriate breaks
}
\description{
Helper function to reproduce scanpy stacked violin plot with Seurat.
}
\details{
Credit to: Tang Ming
https://divingintogeneticsandgenomics.rbind.io/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/
}
