% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameRows.R
\name{renameRows}
\alias{renameRows}
\title{Rename Row Names (Ensembl ID to Symbol)}
\usage{
renameRows(df, anno)
}
\arguments{
\item{anno}{Annotation table. Used to search mapping relationships between Ensembl IDs and Symbols.}

\item{data}{A dataframe, rows are features named as Ensembl IDs, columns are cells.}
}
\value{
A dataframe, rows are features renamed as gene Symbols.
}
\description{
For generic analysis of data not generated commercial pipelines (i.e. SMART-SEQ2), we are usually given a raw count matrix,
where rows are Ensembl IDs. To rename the rows from Ensembl IDs to gene Symbols, a few things need to be considered:

First, row names of Seurat object can NOT be modified, so the renaming operation MUST be done on original expression matrix, usually loaded
as a dataframe.
Second, different Ensembl IDs could map to the same gene Symbol, causing duplicated row names, which is forbidden in dataframe. To
circumvent this, we attach a digit to the Symbols.
}
