% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showColorsWithNames.R
\name{showColorsWithNames}
\alias{showColorsWithNames}
\title{showColorsWithNames}
\usage{
showColorsWithNames(colours, label_names = TRUE, cex_label = 1, borders = NULL)
}
\arguments{
\item{colours}{A character vector of colours, with or without names}

\item{label_names}{Whether to show label names}

\item{cex_label}{Size of printed labels, as multiplier of default size.}

\item{borders}{Border color for each tile. Default uses par("fg"). Use border = NA to omit borders.}
}
\description{
A quick and dirty way to show colours with name labels in a plot, slightly modified from show_col() function in `scales` library .
}
