% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{Silhouette}
\alias{Silhouette}
\title{Silhouette Plot}
\usage{
Silhouette(object, cluster_ident, data_from = "pca")
}
\arguments{
\item{object}{Seurat object.}

\item{cluster_ident}{Identity class of clustering.}

\item{data_from}{By default, data_from = "pca", extract pca embedding as input.
In case of data_from = "expression", "data" slot of the expression matrix is used.}
}
\description{
Unsupervised way to choose the optimal clustering resolutions or number of clusters.
Computes distance matrix based on correlation distance and calculate silhouette scores for a given clustering result.
Outputs a silhouette class object and produces a Silhouette plot.
}
