% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{SilhouetteBoot}
\alias{SilhouetteBoot}
\title{Bootstrapped silhouette score}
\usage{
SilhouetteBoot(
  object,
  cluster_ident,
  data_from = "pca",
  balanced = TRUE,
  iteration = 100,
  subsample.size = NULL,
  replace = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{cluster_ident}{Identity class of clustering.}

\item{data_from}{By default, data_from = "pca", extract pca embedding as input data matrix. In case of data_from = "expression", extract data from original expression matrix.}

\item{balanced}{Whether or not to perform balanced subsampling, which selects proportional size of cells from each cluster. Default is TRUE.}

\item{iteration}{Number of iterations for bootstrapping. Default = 100.}

\item{subsample.size}{Sample size for bootstrapping. If unspecified, subsample one fifth (1/5) of the total population.}

\item{replace}{Sampling with or without replacement. Default is TRUE.}

\item{verbose}{Display progress information.}
}
\description{
Unsupervised way to choose the optimal clustering resolutions or number of clusters.
Perform bootstrapping (repeated random subsampling with replacement) to achieve better estimation of Silhouette score.
}
