% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSW.R
\name{SSW}
\alias{SSW}
\title{Seurat Standard Workflow (SSW)}
\usage{
SSW(
  object,
  assay = NULL,
  nfeatures = 2000,
  PC_features = NULL,
  npcs = 50,
  dims = 1:50,
  k.param = 20,
  algorithm = 1,
  resolution = 0.3,
  perplexity = 30,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{assay}{"RNA" or "integrated".}

\item{nfeatures}{Number of HVGs selected.}

\item{PC_features}{If supplied, run PCA on these features.}

\item{npcs}{Number of PCs to compute.}

\item{dims}{Chosen dimensionality of the data. Used in SNN-graph construction and Non-linear dimension reduction.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm. Supplied to FindNeighbors.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python. Supplied to Findclusters.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. Supplied to Findclusters.}

\item{perplexity}{Perplexity used in RunTSNE.}

\item{verbose}{Print output.}
}
\description{
Run Seurat Standard Workflow.

- Normalization \cr
- Feature Selection (Highly variable genes, HVGs) \cr
- Data scaling \cr
- PCA \cr
- Construct SNN graph \cr
- Unsupervised Clustering \cr
- Non-linear dimension reduction (tSNE, UMAP) \cr

By default, we compute 50 PCs and use the 1~50 PCs as input dimension. To use an optimized dimensionality parameter, run JackStraw or ElbowPlot
to determine the dimensionality of the dataset and supply to `dims`.
}
\examples{
## pbmc_small is too small, must set perplexity to lower value
pbmc_small <- SSW(pbmc_small, perplexity = 10)

}
