% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedViolin.R
\name{StackedVlnPlot}
\alias{StackedVlnPlot}
\title{Make stacked violin plot}
\usage{
StackedVlnPlot(
  obj,
  features,
  pt.size = 0,
  plot.margin = ggplot2::unit(c(-0.75, 0, -0.75, 0), "cm"),
  wrap_title = NULL,
  break_length = 23,
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{features}{Genes to be plotted.}

\item{pt.size}{Size of jitter points. Default is 0 (No point is plotted).}

\item{plot.margin}{Adjust the white space between each plot.}
}
\value{
A ggplot object
}
\description{
Reproduce scanpy stacked violin plot with Seurat.
}
\details{
Credit to: Tang Ming
https://divingintogeneticsandgenomics.rbind.io/post/stacked-violin-plot-for-visualizing-single-cell-data-in-seurat/
}
